requirements.txt
fastapi[standard]
peewee
bcrypt
python-jose[cryptography]
pydantic-settings

# test
pytest
